feat: add non-regioned context menu components + normalize data slots
